Display images in the image slider
Created imageCounter variable that will be used as a base counter for
the looping of the images.
Created the function appendRemoveImage to display images to the slider.
Called the said function in the setImage function with setInterval to
loop through images with a 5 sec interval.
Set fixed style to the images slider to display images correctly.